requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.3
django-crispy-forms==1.6.0
django-jquery==3.1.0
django-registration==2.1.2
django-registration-redux==1.4
django-toolbelt==0.0.1
djangorestframework==3.5.1
djangorestframework-jwt==1.8.0
gunicorn==19.7.1
Pillow==3.3.0
psycopg2==2.7.1
PyJWT==1.4.2
pytz==2017.2
requests==2.10.0
South==1.0.2
static3==0.7.0
stripe==1.35.0
timeout==0.1.2
whitenoise==3.3.0
